+++ CHANGELOG.md
@@ -1,6 +1,18 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
+## 40.5 - 2022-09-07
+### Added
+- We Are All Made of Stars, We Are All Made of Stars (Heroic) and We Are All Made of Stars (Mythic) to SotFO raid
+- A new pets category in Specials
+
+## 40.4 - 2022-09-07
+### Added
+- A bunch of new German translations
+
+### Fixed
+- Switched the tooltip for Fated SoD and Fated SotFO weekly events
+
 ## 40.3 - 2022-09-04
 ### Fixed
 - Cleaned up the Adjustable Categories section in the options again